requirements.txt
requests>=2.32.0
pytz>=2025.1
Pillow>=12.0.0
flask>=3.0.0
flask-cors>=6.0.0
feedparser>=6.0.11
inky>=2.0.0
playwright>=1.55.0
psutil>=7.0.0
git+https://github.com/hbldh/hitherdither.git
black==26.3.0
ruff>=0.14.0
